+++ lib/new_relic/agent/opentelemetry/trace/span.rb
@@ -32,6 +32,27 @@ module NewRelic
             end
           end
 
+          def add_link(link)
+            return self unless recording?
+
+            # add_span_link lives in OpenTelemetry::AbstractSegmentPatch
+            # and in OpenTelemetry::TransactionPatch
+            finishable&.add_span_link(link)
+            self
+          end
+
+          def add_event(name, attributes: nil, timestamp: nil)
+            # This method adds SpanEvent events to a Span event.
+            # A SpanEvent is used to denote a meaningful, singular point in a
+            # Span's duration.
+            return self unless recording?
+
+            # add_span_event_event lives in OpenTelemetry::AbstractSegmentPatch
+            # and in OpenTelemetry::TransactionPatch
+            finishable&.add_span_event_event(name, attributes: attributes, timestamp: timestamp)
+            self
+          end
+
           def apply_translated_attributes(translated)
             translated[:instance_variable].each do |key, value|
               sym_key = "@#{key}".to_sym